вынес классы в отдельные файлы, поправил зависимости; подчистил не используемые сущности.
