+++ src/main/java/Goods.java
@@ -1,0 +1,8 @@
+public class Goods {// Описывает позицию в счете
+    String name;
+    double price;
+    public Goods(String name_, double price_) {
+        name = name_;
+        price = price_;
+    }
+}
